Close optimization for Rows
When a query returns direct results it is possible that all the query result rows are contained in the direct response.  In this case the server can automatically close the operation.  This is indicated by a non-null CloseOperation member in the TSparkDirectResults structure.
In this case calls to Rows.Close don't require any interaction with the server as the operation is already closed.
- added a NewRows function and updated connection to use it when creating Rows
- added logic in NewRows to check for an already closed operation
- updated Rows.Close so we don't make a server round trip when not necessary

Signed-off-by: Raymond Cypher <raymond.cypher@databricks.com>