fix thread safety issue (#74)
Stack created with [Sapling](https://sapling-scm.com). Best reviewed
with
[ReviewStack](https://reviewstack.dev/databricks/databricks-sql-go/pull/74).
* __->__ #74

fix thread safety issue

This fixes a thread safety (and structural issue) with connector.
Because connectors have a long lifetime (potentially process-lifetime),
and are used by multiple threads it's not safe to share the same thrift
client connection.

This changes the connector so that a new thrift client is created on
each connection. The go DB library will perform pooling of connections,
so we don't need to worry about it here.

Signed-off-by: Carl Verge <carl.verge@databricks.com>
Co-authored-by: Carl Verge <carlverge@gmail.com>