Feat: Support ability to retry on specific failures (#95)
# Summary

## Retryable client
In this commit we use RetryableClient to allow for the driver to retry
in case of failure.

Some refactor of authentication to allow for further customization.

Don't retry if the error was due to too many redirects.
Don't retry if the error was due to an invalid protocol scheme.
Don't retry if the error was due to TLS cert verification failure.
Else: retry as it is probably recoverable

TooManyRequests (429) or ServiceUnavailable (503)
If the client receives 503 or 429 with Retry-After header, it should
retry before throwing an error for the user, waiting the amount of
seconds specified by Retry-After header between retries.

## Error handling
Retrieval of specific Databricks headers to enrich errors.

`X-Databricks-Org-Id`
`X-Databricks-Reason-Phrase`
`X-Thriftserver-Error-Message`
`X-Databricks-Error-Or-Redirect-Message`

## Changes in Public facing APIs

To allow for retry configuration, added a new option `WithRetries`:

WithRetries sets up retrying logic. Sane defaults are provided. Negative
retryMax will disable retry behavior
By default retryWaitMin = 1 * time.Second
By default retryWaitMax = 30 * time.Second
By default retryMax = 4

## Testing
Added ability to create end to end tests that simulate retry scenarios

Signed-off-by: Andre Furlan <andre.furlan@databricks.com>