Public error types
Refactored so that the interfaces for system, request, and execution errors are public in dbsql/errors.
Updated error implementations to work better with errors.Is() and errors.As().
Added sentinel values (SystemFault, RequestError, ExecutionError) in dbsql/errors to use with errors.Is()
Updated doc.go to include error information.
Updated errors examples.

Signed-off-by: Raymond Cypher <raymond.cypher@databricks.com>