Retry FetchResults and error handling
Allow retries on calls to fetch results.  The drivers handling of fetching result pages is robust enough to handle unexpected pagination due to retries.
Updated the error handling to make sure that the client error handler will be called so that information in databricks response headers will be added to the error info.
Signed-off-by: Raymond Cypher <raymond.cypher@databricks.com>