Improve error information when query terminates in unexpected state.
When a query terminated in either an unknown or an unexpected operation state the connection was creating a driver error with a static message.
Updated to include the operation state in the error message for an unknown operation state situation.
Updated the case of an unexpected operation state to include the state and the operation status display message.
Updated unit tests based on the new error message.
Signed-off-by: Raymond Cypher <raymond.cypher@databricks.com>