Bug fix issue 147: BUG with reading table that contains copied map
Issue with reading a table with a MAP column, but only when the table was created by copying an existing table.
The bug was caused by an indexing issue when reading map values from an arrow batch.  The occurrence of the bug was dependent on the configuration of batches the result set was broken down into.
Added a new test using the arrow batch returned in the error scenario.  Also fixed an existing test that had been masking this bug.

Signed-off-by: Raymond Cypher <raymond.cypher@databricks.com>