Added a new error type that identifies as driver.ErrBadConn.  This can be added in to an error chain to signal that a connection should no longer be used and to retry with a new connection.
Updated ThriftServiceClient error handling. When a Thrift request fails we are now checking the error message for the term 'Invalid SessionHandle'.  If it is present we are adding a bad connection error to the error chain.  Searching in the error message is clumsy and fragile but there doesn't currently appear to be another way to get the information.

Updated the WithServerHostname function to handle host names prefixed by 'http:' or 'https:' to allow users to specify which protocol to use.  This is in response to github issue #140.

Signed-off-by: Raymond Cypher <raymond.cypher@databricks.com>