PECO-1054 Expose Arrow batches to users
Step one of exposing arrow batches directly to users.
Moved the logic for iterating of the pages in a result set into ResultPageEditor.  Rows now composes in ResultPageIterator.
Introduced Delimeter type. Delimeter tracks a start/end point and provides functions for determining if a point is within the delimiter range and the direction of the point if it is outside the delimeter range.
Updated sparArrowBatch, arrowRowScanner, columnRows, rows to use Delimiter.
Updated the Fetch logic for cloudURL and localBatch so that the concurrentFetcher doesn't need to hold or pass through a Config instance.

Signed-off-by: Raymond Cypher <raymond.cypher@databricks.com>