Bug fix: The HasNext() function for Go driver arrow record iterator returning EOF for empty resultset.
Essentially, the HasNext() logic was assuming there were records in the result set in the case that the original return from executing the statement didn't specify, or if there were direct results.

Updated arrowRecordIterator.HasNext() to check state of underlying iterators. This handles the case of direct results that contain no records.
Updated resultPageIterator.HasNext() to try fetching a result page if necessary to determine if there are more records.

Added tests for empty result sets with direct results enabled and disabled.

Signed-off-by: Raymond Cypher <raymond.cypher@databricks.com>